load and validate inside DappletContext
