+++ src/types/dappletFrame.ts
@@ -1,4 +1,7 @@
+import { DappletTemplate } from './dappletTemplate';
+
 export type DappletFrame = {
     dappletId: string;
+    dapplet?: DappletTemplate;
     txMeta?: any;
 }
